[REFACTORING] Relocate Keywords in mail package
