+++ server/protocols/jmap-rfc-8621/src/main/scala/org/apache/james/jmap/utils/quotas/QuotaReader.scala
@@ -21,9 +21,9 @@ package org.apache.james.jmap.utils.quotas
 
 import javax.inject.Inject
 import org.apache.james.core.quota.{QuotaLimitValue, QuotaUsageValue}
+import org.apache.james.jmap.core.UnsignedInt
+import org.apache.james.jmap.core.UnsignedInt.UnsignedInt
 import org.apache.james.jmap.mail._
-import org.apache.james.jmap.model.UnsignedInt
-import org.apache.james.jmap.model.UnsignedInt.UnsignedInt
 import org.apache.james.mailbox.exception.MailboxException
 import org.apache.james.mailbox.model.{Quota => ModelQuota}
 import org.apache.james.mailbox.quota.QuotaManager
