+++ server/protocols/jmap-rfc-8621/src/test/scala/org/apache/james/jmap/core/KeywordsTest.scala
@@ -22,8 +22,8 @@ package org.apache.james.jmap.core
 import javax.mail.Flags
 import javax.mail.Flags.Flag
 import nl.jqno.equalsverifier.EqualsVerifier
-import org.apache.james.jmap.mail.Keywords
 import org.apache.james.jmap.mail.KeywordsFactory.{LENIENT_KEYWORDS_FACTORY, STRICT_KEYWORDS_FACTORY}
+import org.apache.james.jmap.mail.{Keyword, Keywords}
 import org.apache.james.mailbox.FlagsBuilder
 import org.assertj.core.api.Assertions.assertThat
 import org.scalatest.matchers.must.Matchers
